Test registration in terms of public API
